@@ -4,13 +4,14 @@
   "repository": "github:kopia/kopia",
   "dependencies": {
     "auto-launch": "^5.0.6",
-    "electron-log": "^5.1.2",
-    "electron-store": "^8.2.0",
-    "electron-updater": "^6.3.0-alpha.6",
+    "electron-log": "^5.1.5",
+    "electron-store": "^10.0.0",
+    "electron-updater": "^6.3.0-alpha.7",
     "minimist": "^1.2.8",
     "semver": "^7.6.0",
-    "uuid": "^9.0.1"
+    "uuid": "^10.0.0"
   },
+  "type": "module",
   "author": {
     "email": "kopia-pmc@googlegroups.com",
     "name": "Kopia Authors"
@@ -70,7 +71,7 @@
           ]
         }
       ],
-      "sign": "./sign.js",
+      "sign": "./sign.cjs",
       "signingHashAlgorithms": [
         "sha256"
       ]
@@ -107,17 +108,17 @@
         }
       ]
     },
-    "afterSign": "notarize.js"
+    "afterSign": "notarize.cjs"
   },
   "devDependencies": {
-    "@electron/notarize": "^2.3.0",
-    "@playwright/test": "^1.42.1",
+    "@electron/notarize": "^2.3.2",
+    "@playwright/test": "^1.45.1",
     "asar": "^3.2.0",
     "concurrently": "^8.2.2",
     "dotenv": "^16.4.5",
-    "electron": "^29.1.6",
+    "electron": "^31.2.0",
     "electron-builder": "^24.13.3",
-    "electron-store": "^8.2.0",
+    "electron-store": "^10.0.0",
     "playwright": "^1.37.1",
     "playwright-core": "^1.35.1"
   },
