@@ -1,7 +1,8 @@
-const { app } = require('electron');
-const path = require('path');
+import { app } from 'electron';
+import path from 'path';
+const __dirname = import.meta.dirname;
 
-const osShortName = function() {
+const osShortName = function () {
     switch (process.platform) {
         case "win32":
             return "win"
@@ -14,29 +15,37 @@ const osShortName = function() {
     }
 }();
 
-module.exports = {
-    resourcesPath: function () {
-        if (!app.isPackaged) {
-            return path.join(__dirname, "..", "resources", osShortName);
-        }
-        return process.resourcesPath;
-    },
-    defaultServerBinary: function () {
-        if (!app.isPackaged) {
-            return {
-                "mac": path.join(__dirname, "..", "..", "dist", "kopia_darwin_amd64", "kopia"),
-                "win": path.join(__dirname, "..", "..", "dist", "kopia_windows_amd64", "kopia.exe"),
-                "linux": path.join(__dirname, "..", "..", "dist", "kopia_linux_amd64", "kopia"),
-            }[osShortName]
-        }
+export function iconsPath() {
+    if (!app.isPackaged) {
+        return path.join(__dirname, "..", "resources", osShortName, "icons");
+    }
 
+    return path.join(process.resourcesPath, "icons");
+}
+
+export function publicPath() {
+    if (!app.isPackaged) {
+        return path.join(__dirname, "..", "public");
+    }
+
+    return process.resourcesPath;
+}
+
+export function defaultServerBinary() {
+    if (!app.isPackaged) {
         return {
-            "mac": path.join(process.resourcesPath, "server", "kopia"),
-            "win": path.join(process.resourcesPath, "server", "kopia.exe"),
-            "linux": path.join(process.resourcesPath, "server", "kopia"),
+            "mac": path.join(__dirname, "..", "..", "dist", "kopia_darwin_amd64", "kopia"),
+            "win": path.join(__dirname, "..", "..", "dist", "kopia_windows_amd64", "kopia.exe"),
+            "linux": path.join(__dirname, "..", "..", "dist", "kopia_linux_amd64", "kopia"),
         }[osShortName]
-    },
-    selectByOS: function (x) {
-        return x[osShortName]
-    },
+    }
+
+    return {
+        "mac": path.join(process.resourcesPath, "server", "kopia"),
+        "win": path.join(process.resourcesPath, "server", "kopia.exe"),
+        "linux": path.join(process.resourcesPath, "server", "kopia"),
+    }[osShortName]
+}
+export function selectByOS(x) {
+    return x[osShortName]
 }
